copied all cache tests over
